+++ minecraft/src/net/minecraft/client/model/QuadrupedModel.java
@@ -20,22 +20,22 @@ public class QuadrupedModel<T extends Entity> extends EntityModel<T> {
 
 	public QuadrupedModel(int i, float f) {
 		this.head = new ModelPart(this, 0, 0);
-		this.head.addBox(-4.0F, -4.0F, -8.0F, 8, 8, 8, f);
+		this.head.addBox(-4.0F, -4.0F, -8.0F, 8.0F, 8.0F, 8.0F, f);
 		this.head.setPos(0.0F, (float)(18 - i), -6.0F);
 		this.body = new ModelPart(this, 28, 8);
-		this.body.addBox(-5.0F, -10.0F, -7.0F, 10, 16, 8, f);
+		this.body.addBox(-5.0F, -10.0F, -7.0F, 10.0F, 16.0F, 8.0F, f);
 		this.body.setPos(0.0F, (float)(17 - i), 2.0F);
 		this.leg0 = new ModelPart(this, 0, 16);
-		this.leg0.addBox(-2.0F, 0.0F, -2.0F, 4, i, 4, f);
+		this.leg0.addBox(-2.0F, 0.0F, -2.0F, 4.0F, (float)i, 4.0F, f);
 		this.leg0.setPos(-3.0F, (float)(24 - i), 7.0F);
 		this.leg1 = new ModelPart(this, 0, 16);
-		this.leg1.addBox(-2.0F, 0.0F, -2.0F, 4, i, 4, f);
+		this.leg1.addBox(-2.0F, 0.0F, -2.0F, 4.0F, (float)i, 4.0F, f);
 		this.leg1.setPos(3.0F, (float)(24 - i), 7.0F);
 		this.leg2 = new ModelPart(this, 0, 16);
-		this.leg2.addBox(-2.0F, 0.0F, -2.0F, 4, i, 4, f);
+		this.leg2.addBox(-2.0F, 0.0F, -2.0F, 4.0F, (float)i, 4.0F, f);
 		this.leg2.setPos(-3.0F, (float)(24 - i), -5.0F);
 		this.leg3 = new ModelPart(this, 0, 16);
-		this.leg3.addBox(-2.0F, 0.0F, -2.0F, 4, i, 4, f);
+		this.leg3.addBox(-2.0F, 0.0F, -2.0F, 4.0F, (float)i, 4.0F, f);
 		this.leg3.setPos(3.0F, (float)(24 - i), -5.0F);
 	}
