+++ minecraft/src/net/minecraft/client/model/SilverfishModel.java
@@ -20,7 +20,8 @@ public class SilverfishModel<T extends Entity> extends EntityModel<T> {
 
 		for (int i = 0; i < this.bodyParts.length; i++) {
 			this.bodyParts[i] = new ModelPart(this, BODY_TEXS[i][0], BODY_TEXS[i][1]);
-			this.bodyParts[i].addBox((float)BODY_SIZES[i][0] * -0.5F, 0.0F, (float)BODY_SIZES[i][2] * -0.5F, BODY_SIZES[i][0], BODY_SIZES[i][1], BODY_SIZES[i][2]);
+			this.bodyParts[i]
+				.addBox((float)BODY_SIZES[i][0] * -0.5F, 0.0F, (float)BODY_SIZES[i][2] * -0.5F, (float)BODY_SIZES[i][0], (float)BODY_SIZES[i][1], (float)BODY_SIZES[i][2]);
 			this.bodyParts[i].setPos(0.0F, (float)(24 - BODY_SIZES[i][1]), f);
 			this.zPlacement[i] = f;
 			if (i < this.bodyParts.length - 1) {
@@ -30,13 +31,13 @@ public class SilverfishModel<T extends Entity> extends EntityModel<T> {
 
 		this.bodyLayers = new ModelPart[3];
 		this.bodyLayers[0] = new ModelPart(this, 20, 0);
-		this.bodyLayers[0].addBox(-5.0F, 0.0F, (float)BODY_SIZES[2][2] * -0.5F, 10, 8, BODY_SIZES[2][2]);
+		this.bodyLayers[0].addBox(-5.0F, 0.0F, (float)BODY_SIZES[2][2] * -0.5F, 10.0F, 8.0F, (float)BODY_SIZES[2][2]);
 		this.bodyLayers[0].setPos(0.0F, 16.0F, this.zPlacement[2]);
 		this.bodyLayers[1] = new ModelPart(this, 20, 11);
-		this.bodyLayers[1].addBox(-3.0F, 0.0F, (float)BODY_SIZES[4][2] * -0.5F, 6, 4, BODY_SIZES[4][2]);
+		this.bodyLayers[1].addBox(-3.0F, 0.0F, (float)BODY_SIZES[4][2] * -0.5F, 6.0F, 4.0F, (float)BODY_SIZES[4][2]);
 		this.bodyLayers[1].setPos(0.0F, 20.0F, this.zPlacement[4]);
 		this.bodyLayers[2] = new ModelPart(this, 20, 18);
-		this.bodyLayers[2].addBox(-3.0F, 0.0F, (float)BODY_SIZES[4][2] * -0.5F, 6, 5, BODY_SIZES[1][2]);
+		this.bodyLayers[2].addBox(-3.0F, 0.0F, (float)BODY_SIZES[4][2] * -0.5F, 6.0F, 5.0F, (float)BODY_SIZES[1][2]);
 		this.bodyLayers[2].setPos(0.0F, 19.0F, this.zPlacement[1]);
 	}
